Use beacon service to send impression request when an ad is requested [#63842932]

+++ SharethroughSDK/Extension Helpers/STRDateProvider.h
@@ -11,6 +11,6 @@
 @interface STRDateProvider : NSObject
 
 - (NSDate *)now;
-- (long)millisecondsSince1970;
+- (long long)millisecondsSince1970;
 
 @end
